+++ Reads/Controllers/CategoriesController.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
-using Reads;
 using Reads.Models;
 
 namespace Reads.Controllers
@@ -9,19 +8,19 @@ namespace Reads.Controllers
     [ApiController]
     public class CategoriesController : ControllerBase
     {
-        private readonly ICategoryRepository _categoryRepository;
+        private readonly IEfRepository<Category> _categoryRepository;
 
-        public CategoriesController(ICategoryRepository categoryRepository)
+        public CategoriesController(IEfRepository<Category> categoryRepository)
         {
             _categoryRepository = categoryRepository;
         }
 
         // GET api/categories
         [HttpGet]
-        public ActionResult<IEnumerable<string>> Get()
+        public ActionResult<IEnumerable<Category>> Get()
         {
             List<Category> categories = _categoryRepository.GetAll().Result;
-            return Ok(categories);
+            return categories;
         }
 
         // POST api/categories
@@ -57,6 +56,10 @@ namespace Reads.Controllers
                 return BadRequest(ModelState);
             }
             var category = _categoryRepository.Get(id).Result;
+            if (category == null)
+            {
+                return BadRequest($"No category found for id: {id}");
+            }
             _categoryRepository.Delete(category);
             return Ok();
         }
